+++ sdk/ai/azure-ai-agents/src/main/java/com/azure/ai/agents/BetaAgentEndpointConversationsClient.java
@@ -304,7 +304,7 @@ public final class BetaAgentEndpointConversationsClient {
      *     max_output_tokens: BinaryData (Optional)
      *     output (Optional): [
      *          (Optional){
-     *             type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *             type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      *         }
      *     ]
      *     audio (Optional): {
@@ -391,7 +391,7 @@ public final class BetaAgentEndpointConversationsClient {
      *     max_output_tokens: BinaryData (Optional)
      *     output (Optional): [
      *          (Optional){
-     *             type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *             type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      *         }
      *     ]
      *     audio (Optional): {
@@ -467,7 +467,7 @@ public final class BetaAgentEndpointConversationsClient {
      * <pre>
      * {@code
      * {
-     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      * }
      * }
      * </pre>
@@ -521,7 +521,7 @@ public final class BetaAgentEndpointConversationsClient {
      * <pre>
      * {@code
      * {
-     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      * }
      * }
      * </pre>
@@ -555,7 +555,7 @@ public final class BetaAgentEndpointConversationsClient {
      * <pre>
      * {@code
      * {
-     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      * }
      * }
      * </pre>