+++ sdk/ai/azure-ai-agents/src/main/java/com/azure/ai/agents/implementation/BetaAgentEndpointConversationsImpl.java
@@ -983,7 +983,7 @@ public final class BetaAgentEndpointConversationsImpl {
      *     max_output_tokens: BinaryData (Optional)
      *     output (Optional): [
      *          (Optional){
-     *             type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *             type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      *         }
      *     ]
      *     audio (Optional): {
@@ -1096,7 +1096,7 @@ public final class BetaAgentEndpointConversationsImpl {
      *     max_output_tokens: BinaryData (Optional)
      *     output (Optional): [
      *          (Optional){
-     *             type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *             type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      *         }
      *     ]
      *     audio (Optional): {
@@ -1204,7 +1204,7 @@ public final class BetaAgentEndpointConversationsImpl {
      *     max_output_tokens: BinaryData (Optional)
      *     output (Optional): [
      *          (Optional){
-     *             type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *             type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      *         }
      *     ]
      *     audio (Optional): {
@@ -1315,7 +1315,7 @@ public final class BetaAgentEndpointConversationsImpl {
      *     max_output_tokens: BinaryData (Optional)
      *     output (Optional): [
      *          (Optional){
-     *             type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *             type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      *         }
      *     ]
      *     audio (Optional): {
@@ -1402,7 +1402,7 @@ public final class BetaAgentEndpointConversationsImpl {
      *     max_output_tokens: BinaryData (Optional)
      *     output (Optional): [
      *          (Optional){
-     *             type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *             type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      *         }
      *     ]
      *     audio (Optional): {
@@ -1495,7 +1495,7 @@ public final class BetaAgentEndpointConversationsImpl {
      *     max_output_tokens: BinaryData (Optional)
      *     output (Optional): [
      *          (Optional){
-     *             type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *             type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      *         }
      *     ]
      *     audio (Optional): {
@@ -1571,7 +1571,7 @@ public final class BetaAgentEndpointConversationsImpl {
      * <pre>
      * {@code
      * {
-     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      * }
      * }
      * </pre>
@@ -1631,7 +1631,7 @@ public final class BetaAgentEndpointConversationsImpl {
      * <pre>
      * {@code
      * {
-     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      * }
      * }
      * </pre>
@@ -1685,7 +1685,7 @@ public final class BetaAgentEndpointConversationsImpl {
      * <pre>
      * {@code
      * {
-     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      * }
      * }
      * </pre>
@@ -1743,7 +1743,7 @@ public final class BetaAgentEndpointConversationsImpl {
      * <pre>
      * {@code
      * {
-     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      * }
      * }
      * </pre>
@@ -1796,7 +1796,7 @@ public final class BetaAgentEndpointConversationsImpl {
      * <pre>
      * {@code
      * {
-     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      * }
      * }
      * </pre>
@@ -1853,7 +1853,7 @@ public final class BetaAgentEndpointConversationsImpl {
      * <pre>
      * {@code
      * {
-     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      * }
      * }
      * </pre>
@@ -1905,7 +1905,7 @@ public final class BetaAgentEndpointConversationsImpl {
      * <pre>
      * {@code
      * {
-     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      * }
      * }
      * </pre>
@@ -1960,7 +1960,7 @@ public final class BetaAgentEndpointConversationsImpl {
      * <pre>
      * {@code
      * {
-     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      * }
      * }
      * </pre>
@@ -1994,7 +1994,7 @@ public final class BetaAgentEndpointConversationsImpl {
      * <pre>
      * {@code
      * {
-     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      * }
      * }
      * </pre>
@@ -2033,7 +2033,7 @@ public final class BetaAgentEndpointConversationsImpl {
      * <pre>
      * {@code
      * {
-     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request) (Required)
+     *     type: String(function_call/function_call_output/mcp_approval_response/mcp_list_tools/mcp_call/mcp_approval_request/message) (Required)
      * }
      * }
      * </pre>